+++ packages/nonebot-adapter-cqhttp/nonebot/adapters/cqhttp/bot.py
@@ -396,7 +396,7 @@ class Bot(BaseBot):
           - ``NetworkError``: 网络错误
           - ``ActionFailed``: API 调用失败
         """
-        return super().call_api(api, **data)
+        return await super().call_api(api, **data)
 
     @overrides(BaseBot)
     async def send(self,